CReated a new project card and added a new role attribute to profiles. This allows us to differentiate the two parts of the page for the project section for admin viewing
